+++ Changelog.md
@@ -7,6 +7,7 @@
     - Window-World : Rename and move WorldCallbacks to WindowCallbacks
     - WindowCallbacks : Add CHANGEWORLD and STOPWINDOW
     - Entity : Can't have the same type of component two times
+    - EntitySystem-UISystem : Change debug color to blue
     - PhysicsComponent : Create CollisionInfos
     - MoveComponent : Unlock diagonal movement
     - LifeComponent : Use clamp function to set life